Refactor use-ticket-info and remove Root_CarnetDetailsScreen (#4251)
* refactor: remove navigation from use-ticket-info hook and removed root carnet ticket screen

* refcator: remove unnecessary type and use farecontract to determine ticket type is carnet or not

* refactor: move function to utils

* refactor: fix import
