+++ src/components/screen-header/use-ticket-info.tsx
@@ -1,16 +1,18 @@
-import {useAnalytics} from '@atb/analytics';
 import {
-    PreassignedFareProduct,
+  PreassignedFareProduct,
   findReferenceDataById,
   isOfFareProductRef,
   useFirestoreConfiguration,
 } from '@atb/configuration';
-import {RootNavigationProps} from '@atb/stacks-hierarchy';
 import {FareContract, useTicketingState} from '@atb/ticketing';
-import {useNavigation} from '@react-navigation/native';
+
+type TicketInfoParams = {
+  fareProductTypeConfigType: string;
+  preassignedFareProductId: string;
+};
 
 type TicketInfo = {
-  navigateToTicketInfoScreen: () => void;
+  ticketInfoParams: TicketInfoParams | undefined;
   fareContract: FareContract | undefined;
   preassignedFareProduct: PreassignedFareProduct | undefined;
 };
@@ -26,28 +28,14 @@ export const useTicketInfo = (orderId: string): TicketInfo => {
     isOfFareProductRef(firstTravelRight) ? firstTravelRight.fareProductRef : '',
   );
 
-  const analytics = useAnalytics();
-  const navigation = useNavigation<RootNavigationProps>();
-
   const ticketInfoParams = preassignedFareProduct && {
     fareProductTypeConfigType: preassignedFareProduct?.type,
     preassignedFareProductId: preassignedFareProduct?.id,
   };
 
-  const navigateToTicketInfoScreen = () => {
-    ticketInfoParams &&
-      analytics.logEvent(
-        'Ticketing',
-        'Ticket information button clicked',
-        ticketInfoParams,
-      );
-    ticketInfoParams &&
-      navigation.navigate('Root_TicketInformationScreen', ticketInfoParams);
-  };
-
   return {
-    navigateToTicketInfoScreen,
+    ticketInfoParams,
     fareContract,
-    preassignedFareProduct
+    preassignedFareProduct,
   };
 };
