+++ src/fare-contracts/FareContractAndReservationsList.tsx
@@ -74,7 +74,9 @@ export const FareContractAndReservationsList: React.FC<Props> = ({
               analytics.logEvent('Ticketing', 'Ticket details clicked');
               navigation.navigate({
                 name: 'Root_FareContractDetailsScreen',
-                params: {orderId: fcOrReservation.orderId},
+                params: {
+                  orderId: fcOrReservation.orderId,
+                },
               });
             }}
             key={fcOrReservation.orderId}
